Migrate subprojects config. to script plugin
